Implement semantic exception hierarchy with proper inheritance patterns
